+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/api/data/DataApiController.java
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.f12.notionlinkedblog.api.common.Endpoint;
-import io.f12.notionlinkedblog.exceptions.ExceptionMessages;
+import io.f12.notionlinkedblog.exceptions.message.ExceptionMessages;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
