+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/post/PostService.java
@@ -1,8 +1,8 @@
 package io.f12.notionlinkedblog.service.post;
 
-import static io.f12.notionlinkedblog.exceptions.ExceptionMessages.PostExceptionsMessages.IMAGE_NOT_EXIST;
-import static io.f12.notionlinkedblog.exceptions.ExceptionMessages.PostExceptionsMessages.*;
-import static io.f12.notionlinkedblog.exceptions.ExceptionMessages.UserExceptionsMessages.*;
+import static io.f12.notionlinkedblog.exceptions.message.ExceptionMessages.PostExceptionsMessages.IMAGE_NOT_EXIST;
+import static io.f12.notionlinkedblog.exceptions.message.ExceptionMessages.PostExceptionsMessages.*;
+import static io.f12.notionlinkedblog.exceptions.message.ExceptionMessages.UserExceptionsMessages.*;
 
 import java.io.File;
 import java.io.IOException;
