+++ 003 Code/BE/notion-linked-blog/src/test/java/io/f12/notionlinkedblog/api/post/PostApiControllerTest.java
@@ -1,6 +1,6 @@
 package io.f12.notionlinkedblog.api.post;
 
-import static io.f12.notionlinkedblog.exceptions.ExceptionMessages.UserExceptionsMessages.*;
+import static io.f12.notionlinkedblog.exceptions.message.ExceptionMessages.UserExceptionsMessages.*;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.http.MediaType.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
