+++ 003 Code/BE/notion-linked-blog/src/test/java/io/f12/notionlinkedblog/service/user/UserServiceTests.java
@@ -1,6 +1,6 @@
 package io.f12.notionlinkedblog.service.user;
 
-import static io.f12.notionlinkedblog.exceptions.ExceptionMessages.UserExceptionsMessages.*;
+import static io.f12.notionlinkedblog.exceptions.message.ExceptionMessages.UserExceptionsMessages.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
